splitted update.php into multiple files for easier maintenance
